requirements.txt
aiofiles==0.5.0
aniso8601==7.0.0
async-exit-stack==1.0.1
async-generator==1.10
attrs==20.3.0
boto3==1.16.55
botocore==1.19.55
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dnspython==2.1.0
email-validator==1.1.2
fastapi==0.63.0
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
h11==0.12.0
httptools==0.1.1
idna==2.10
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
orjson==3.4.6
packaging==20.8
pluggy==0.13.1
promise==2.3
psycopg2-binary==2.8.6
py==1.10.0
pydantic==1.7.3
pyparsing==2.4.7
pytest==6.2.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-multipart==0.0.5
PyYAML==5.3.1
requests==2.25.1
Rx==1.6.1
s3transfer==0.3.4
six==1.15.0
SQLAlchemy==1.3.22
starlette==0.13.6
toml==0.10.2
ujson==3.2.0
urllib3==1.26.2
uvicorn==0.13.3
uvloop==0.14.0
watchgod==0.6
websockets==8.1
